Organize routes under versioned folders
