+++ src/routes/v1/cultures.routes.js
@@ -1,6 +1,6 @@
 const express = require("express");
 const router = express.Router();
-const culturesController = require("../controllers/cultures.controller");
+const culturesController = require("../../controllers/cultures.controller");
 
 router.get("/", culturesController.getAllCultures);
 router.get("/years", culturesController.getAvailableYears);
