+++ src/routes/v1/import.routes.js
@@ -1,7 +1,7 @@
 const express = require("express");
 const multer = require("multer");
 const router = express.Router();
-const importController = require("../controllers/import.controller");
+const importController = require("../../controllers/import.controller");
 
 const storage = multer.memoryStorage();
 const upload = multer({ storage });
