+++ src/routes/v1/regions.routes.js
@@ -1,6 +1,6 @@
 const express = require("express");
 const router = express.Router();
-const regionsController = require("../controllers/regions.controller");
+const regionsController = require("../../controllers/regions.controller");
 
 router.get("/", regionsController.getAllRegions);
