+++ src/routes/v1/stats.routes.js
@@ -1,7 +1,7 @@
 const express = require("express");
 const router = express.Router();
-const statsController = require("../controllers/stats.controller");
-const geoController = require('../controllers/geo.controller');
+const statsController = require("../../controllers/stats.controller");
+const geoController = require('../../controllers/geo.controller');
 
 router.get("/regions/cultures/:cultureId/years/:year", statsController.getStatsByRegion);
 router.get("/cultures/:id/summary", statsController.getCultureSummary);
